ai-service/requirements.txt
# Core Framework
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# AI & Computer Vision
openai>=1.50.0
httpx>=0.27.0
opencv-python-headless>=4.8.1.78
pytesseract>=0.3.10
pillow>=10.1.0
numpy>=1.26.2

# Background Removal (and its dependencies)
rembg>=2.0.50
onnxruntime>=1.16.0

# Data Validation
pydantic>=2.5.2